fix(loader): avoid uint16 lshift in ue8m0->bf16 conversion
PyTorch CPU has no lshift kernel for UInt16, so the previous
`(scale_t.to(torch.uint16) << 7)` raised NotImplementedError when
loading any V4-Flash MXFP4 routed-expert scale tensor on the host.

Switch to int32 for the shift (kernel exists) and narrow to int16
afterwards. The shifted value max is 255<<7 = 32640, well within
int16 range, so the narrow is lossless. The .view(bfloat16) bit
pattern is identical (bf16 sign bit is always 0 for ue8m0 values).

+++ kt-kernel/python/utils/loader.py
@@ -1264,7 +1264,9 @@ class MXFP4SafeTensorLoader(SafeTensorLoader):
         # bf16 = [sign(1) | exp(8) | mant(7)]; setting mant=0, exp=e gives 2^(e-127),
         # which is exactly the value encoded by ue8m0 for e ∈ [1, 254]. e=0 → bf16 +0
         # (acceptable: ue8m0=0 represents 2^-127, below bf16 normal range), e=255 → +inf.
-        return (scale_t.to(torch.uint16) << 7).view(torch.bfloat16).contiguous()
+        # Compute in int32 then narrow to int16 (max value is 255<<7=32640, fits int16),
+        # because torch CPU has no lshift kernel for uint16.
+        return (scale_t.to(torch.int32) << 7).to(torch.int16).view(torch.bfloat16).contiguous()
 
     def load_experts(self, base_key: str, device: str = "cpu"):
         gate_name, up_name, down_name = self.PROJ_NAMES
